Fixed bug where SAS files wouldn't decode to UTF-8 with bad value
After much hair pulling, I finalized realized why the SQLite database wasn't displaying the text without showing up as binary: Because some entries (~40 in the main kidpan database) had bytes that did not decode in utf-8. After briefly considering recompiling python 3.9 with the desired handling (which while it sounded fun, may not have been the most practical approach to this solution), but ultimately landed on the 'replace' function of the errors argument in the decode function. The detail is to raise an error (Unicode Error or something like that), which is why my code kept failing. In this case, it replaces bytes that can't be decoded to 0xFFFD, or the Unicode Replacement Character. It is easily enough search through (i.e. with grep) to filter these rows. A back of the envelope calculation shows that 0.004% of all rows in the main table (kidpan) have this, so it won't be missed